pyproject.toml
[project]
name = "sca"
version = "0.0.1"
description = "Spy Cats Agency!"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "django-cors-headers>=4.9.0",
    "django-split-settings>=1.3.2",
    "django-unfold>=0.66.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "drf-standardized-errors>=0.15.0",
    "pillow>=11.3.0",
    "python-decouple>=3.8",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "isort>=6.0.1",
]
